Add CLI servoterm and faithful scope protocol to host sim
Port the 'term' component into the host build via a CDC shim that maps the
device's USB-serial byte stream onto the process stdin/stdout. A new --serve
mode speaks the exact wire protocol (text responses interleaved with 0xFF
scope frames) and drains scope continuously during a run, staying
deterministic since time advances only via run/step.

Add servoterm (host/servoterm.c): a headless QtServoterm-style client that
spawns the simulator, forwards commands, and demuxes the reply stream into
text plus a decoded scope CSV, for building automated test assets.

Widen three scanf locals in hal.c from int32_t to long: the firmware assumes
a 32-bit long (ARM ILP32), but on an LP64 host "%li"/"%lu" wrote 8 bytes into
a 4-byte slot and corrupted the parse. No effect on the 32-bit target.

+++ shared/hal.c
@@ -799,13 +799,13 @@ uint32_t hal_parse_(char *cmd) {
   int32_t foo = 0;
 
   char sinkc[64];
-  int32_t sinki = 0;
+  long sinki = 0;  // long (not int32_t) so the %li scanf width matches on LP64 hosts too
   char sinkp[64];
 
   float value = 0.0;
 
   char sourcec[64];
-  int32_t sourcei = 0;
+  long sourcei = 0;  // see sinki
   char sourcep[64];
 
   hal_pin_inst_t *sink;
@@ -967,7 +967,7 @@ COMMAND("debug_level", debug_level, "set hal debug level, 0 = print all, 1 = pri
 
 void hal_linked_pins(char *ptr){
   char sinkc[64];
-  uint32_t sinki = 0;
+  unsigned long sinki = 0;  // unsigned long so the %lu scanf width matches on LP64 hosts too
   char sinkp[64];
   int foo;
   foo = sscanf(ptr, " %[a-zA-Z_]%lu.%[a-zA-Z0-9_]", sinkc, &sinki, sinkp);
